// packages/database/src/pg/models/airdrop_update.ts
/**
 * This file was automatically generated by knex
 * Do not modify this file manually
 */
import { Increment } from '../interfaces/common';
import * as commons from './common_update';

export interface PgAirdropUpdate extends commons.BaseRecordUpdate {
  member?: string | null;
  token?: string | null;
  award?: string | null;
  vestingAt?: Date | null;
  count?: number | null | Increment;
  status?: string | null;
  orderId?: string | null;
  billPaymentId?: string | null;
  sourceAddress?: string | null;
  stakeRewardId?: string | null;
  stakeType?: string | null;
  isBaseToken?: boolean | null;
}